+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 0.6dev
 ------
 
+- Adopt shipping handling to ``bda.plone.shipping`` >= 0.4.
+  [rnix]
+
 - ``CartDataProviderBase`` no longer provides default values for ``currency``,
   ``cart_url``, ``show_to_cart``, ``show_checkout`` and ``show_currency``.
   [rnix]